requirements.txt
streamlit
langchain
langchain-community
sentence-transformers
pypdf
openai
chromadb
git+https://github.com/openai/whisper.git
python-dotenv
SpeechRecognition
gtts
streamlit-audiorecorder